Restore default serializer
Without an explicit format (eg ".xml"), the default renderer
for a listing lands up being the Akoma Ntoso XML serializer, which
explicitly doesn't have a serializer class. When we later override
that renderer decision, we also need to bring back in the default
serializer class.

Fixes LAWSAFRICA-1T on Sentry